Create workspace file, job requires admin

+++ jobs/theia.sh
@@ -6,12 +6,35 @@ source $GOPHER_INSTALLDIR/lib/gopherbot_v1.sh
 
 FailTask tail-log
 
-cat > $HOME/.bashrc <<EOF
-# File created by jobs/theia.sh
+if [ ! -e "$HOME/.bashrc" ]
+then
+    cat > $HOME/.bashrc <<EOF
+# File created by jobs/theia.sh; changes will be preserved.
 PS1="\[\033[01;32m\]robot@gopherbot-dev\[\033[00m\]:\[\033[01;34m\]\w\[\033[00m\]\$ "
 PATH=$HOME/bin:$HOME/.local/bin:$HOME/go/bin:/opt/gopherbot:$PATH
 export PATH PS1
 EOF
+fi
+
+if [ ! -e "$HOME/robot.theia-workspace" ]
+then
+    cat > $HOME/robot.theia-workspace <<EOF
+{
+   "folders": [
+      {
+         "path": "file:///home/robot/custom"
+      },
+      {
+         "path": "file:///home/robot/robot-defaults"
+      },
+      {
+         "path": "file:///home"
+      }
+   ],
+   "settings": {}
+}
+EOF
+fi
 
 cat > $HOME/stop-theia.sh <<"EOF"
 kill $PPID
